fix: Fixed incorrect board tiling
The board was set up such that the pattern of light and dark squares
would be different depending on whether the board was flipped or not.
This meant that sometimes the white queen would be on a dark square and
the black queen would be on a light square and we can't have that.